Moved PrintConfig logic to yaml_input.c.
